+++ Engine/Components/Graphic.h
@@ -18,7 +18,7 @@ public:
 	/// Methods (public)
 	virtual Graphic* clone() const override;
 
-	void render(struct RenderContext *ctx, size_t num_views, struct View const *views) const;
+	void render(struct RenderContext *ctx, uint32_t num_views, struct View const *views) const;
 
 	/// Setters
 	void setMesh(MeshRef _mesh);
